+++ src/Controller/TaskController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Task;
+use App\Enum\TaskStatus;
 use App\Form\TaskCreateFormType;
 use App\Repository\TaskRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -10,7 +11,6 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 use Spatie\Browsershot\Browsershot;
-// use Dompdf\Dompdf;
 
 final class TaskController extends AbstractController
 {
@@ -23,7 +23,7 @@ final class TaskController extends AbstractController
     #[Route('/task', name: 'app_task', methods: ['GET'])]
     public function taskList(): Response
     {
-        $taskList = $this->taskRepository->findAll();
+        $taskList = $this->taskRepository->findBy(['status' => TaskStatus::OPEN], ['createdAt' => 'DESC']);
 
         return $this->render('task/index.html.twig', [
             'taskList' => $taskList
@@ -31,7 +31,7 @@ final class TaskController extends AbstractController
     }
 
     #[Route('/task/create', name: 'app_task_create', methods: ['GET', 'POST'])]
-    public function createTaskForm(Request $request): Response
+    public function createTask(Request $request): Response
     {
         $task = new Task();
         $form = $this->createForm(TaskCreateFormType::class, $task)->handleRequest($request);
@@ -46,7 +46,7 @@ final class TaskController extends AbstractController
     }
 
     #[Route('/task/edit/{task}', name: 'app_task_edit', methods: ['GET', 'PATCH'])]
-    public function editTaskForm(Request $request, Task $task): Response
+    public function editTask(Request $request, Task $task): Response
     {
         $form = $this->createForm(TaskCreateFormType::class, $task, ['method' => 'PATCH'])->handleRequest($request);
 
@@ -62,13 +62,51 @@ final class TaskController extends AbstractController
     #[Route('/task/delete/{task}', name: 'app_task_delete', methods: ['DELETE'])]
     public function deleteTask(Request $request, Task $task): Response
     {
-        $this->taskRepository->delete($task, true);
-        $this->addFlash('success', 'Task deleted!');
+        if ($this->isCsrfTokenValid('delete'.$task->getId(), $request->request->get('_token'))) {
+            $this->taskRepository->delete($task, true);
+            $this->addFlash('success', 'Task deleted!');
+        } else {
+            $this->addFlash('danger', 'Security error: Invalid CSRF token.');
+        }
+
         return $this->redirectToRoute('app_task');
     }
 
+    #[Route('task/history', name:'app_task_history', methods: ['GET'])]
+    public function taskHistory(Request $request): Response
+    {
+        $taskList = $this->taskRepository->findBy(['status' => TaskStatus::FINISHED], ['createdAt' => 'DESC']);
+        return $this->render('task/history.html.twig', compact('taskList'));
+    }
+
+    #[Route('/task/finished/{task}', name: 'app_task_finished', methods: ['PATCH'])]
+    public function taskFinished(Request $request, Task $task): Response {
+        if ($task->getFinishedAt() !== null) {
+            $this->addFlash('warning', 'This task is already finished!');
+            return $this->redirectToRoute('app_task');
+        }
+
+        if ($this->isCsrfTokenValid('finished'.$task->getId(), $request->request->get('_token'))) {
+            $task->setFinishedAt(new \DateTimeImmutable());
+            $task->setStatus(TaskStatus::FINISHED);
+            $this->taskRepository->update($task, true);
+            $this->addFlash('success', 'Task finished!');
+        } else {
+            $this->addFlash('danger', 'Security error: Invalid CSRF token.');
+        }
+
+        return $this->redirectToRoute('app_task');
+    }
+
+    #[Route('/task/')]
+
     #[Route('/task/report', name: 'app_task_report', methods: ['GET'])]
-    public function taskReport(Request $request): Response
+    public function taskReport(Request $request): Response {
+        return $this->render('task/report.html.twig', []);
+    }
+
+    #[Route('/task/report/download', name: 'app_task_report_download', methods: ['GET'])]
+    public function taskReportDownload(Request $request): Response
     {
         $taskList = $this->taskRepository->findAll();
         $manifestPath = $this->getParameter('kernel.project_dir') . '/public/build/manifest.json';
@@ -83,7 +121,7 @@ final class TaskController extends AbstractController
                 }
             }
         }
-        $html = $this->renderView('task/report.html.twig', [
+        $html = $this->renderView('task/reportDownload.html.twig', [
             'taskList' => $taskList,
             'inline_css' => $cssContent
         ]);
@@ -93,11 +131,5 @@ final class TaskController extends AbstractController
             'Content-Type' => 'application/pdf',
             'Content-Disposition' => 'attachment; filename="task.pdf"'
         ]);
-//        $dompdf = new Dompdf();
-//        $dompdf->loadHtml($html);
-//        $dompdf->setPaper('A4', 'portrait');
-//        $dompdf->render();
-//        $dompdf->stream("task.pdf", array("Attachment" => true));
-//        exit;
     }
 }
